+++ ios/MemeGTD/MemeGTD/MemeGTDApp.swift
@@ -5,6 +5,7 @@ struct MemeGTDApp: App {
     @StateObject private var taskStore = TaskStore()
     @StateObject private var memoStore = MemoStore()
     @StateObject private var articleStore = ArticleStore()
+    @StateObject private var dataSources = DataSourceProvider()
 
     var body: some Scene {
         WindowGroup {
@@ -12,6 +13,7 @@ struct MemeGTDApp: App {
                 .environmentObject(taskStore)
                 .environmentObject(memoStore)
                 .environmentObject(articleStore)
+                .environmentObject(dataSources)
                 .preferredColorScheme(.light)
         }
     }